fixed GetTime

+++ Tabby/src/Tabby/Core/Timestep.h
@@ -9,18 +9,17 @@ public:
     {
     }
 
-    operator float() const
+    operator double() const
     {
-
         return m_Time;
     }
 
-    float GetSeconds() const { return m_Time; }
-    float GetMilliseconds() const { return m_Time * 1000.0f; }
+    double GetSeconds() const { return m_Time; }
+    double GetMilliseconds() const { return m_Time * 1000.0f; }
 
 private:
     // static Timestep* s_Instance;
-    float m_Time;
+    double m_Time;
 };
 
 }
